Add files via upload
In this repository we have written a code to check whether a number is palindrome is not. A palindrome number is a number which gives the same number after reading it in reverse order for  example 121 , reading it from right to left it gives the same number. another example 234 , reading it from right to left it gives 432 , so it's not a palindrome number